fix hmm_align & add HMM-VAE examples in jupyter
