+++ beer/models/__init__.py
@@ -13,7 +13,7 @@ from .normal import *
 from .mixture import *
 from .hmm import *
 from .phoneloop import *
-#from .vae import *
+from .vae import *
 from .gsm import *
 
 ## Concrete model set.
