+++ beer/nnet/__init__.py
@@ -2,4 +2,4 @@
 from .arnet import *
 from .problayers import *
 from .neuralnetwork import *
-
+from .residual import *
